[bsr][api][finance] Move core.payments.utils function to pcapi.core.finance package
