+++ api/src/pcapi/core/finance/utils.py
@@ -1,7 +1,12 @@
+import datetime
 import decimal
 import typing
 
 from babel import numbers
+import pytz
+
+
+ACCOUNTING_TIMEZONE = pytz.timezone("Europe/Paris")
 
 
 def to_eurocents(amount_in_euros: typing.Union[decimal.Decimal, float]) -> int:
@@ -39,3 +44,13 @@ def format_raw_iban_and_bic(raw_data: str) -> typing.Optional[str]:
     formatted_data = raw_data.upper()
     formatted_data = formatted_data.replace(" ", "")
     return formatted_data
+
+
+def get_cutoff_as_datetime(last_day: datetime.date) -> datetime.datetime:
+    """Return a UTC datetime object for the first second of the day after
+    the requested date (the latter being expressed as a date in the
+    CET/CEST timezone).
+    """
+    next_day = last_day + datetime.timedelta(days=1)
+    first_second = datetime.datetime.combine(next_day, datetime.time.min)
+    return ACCOUNTING_TIMEZONE.localize(first_second).astimezone(pytz.utc)
